Move `image_ref_with_digest` util to `gci.oci` package
